requirement.txt
numpy
langdetect
scikit-learn
sentence-transformers
faiss-cpu
torch
beautifulsoup4
fastapi
uvicorn
pydantic
deep-translator